core: align network stats tx\rx_drops definition & type
rx_drop and tx_drop values reported by vdsm are defined as 'Number of
packets received but not processed', aka -dropped [1][2]. This should be
reflected by their names in engine.
As they are of type U64 (unsigned long) [1], they are whole numbers. This
should be reflected by the type of their counterparts in engine.

[1] https://www.kernel.org/doc/html/latest/networking/
    statistics.html#c.rtnl_link_stats64
[2] https://github.com/oVirt/vdsm/blob/
    51ee47b235ec28744165190a22e211d43fb371b7/lib/vdsm/api/
    vdsm-api.yml#L5076

Bug-Url: https://bugzilla.redhat.com/1980192

+++ frontend/webadmin/modules/uicommonweb/src/main/java/org/ovirt/engine/ui/uicommonweb/models/hosts/HostInterface.java
@@ -88,13 +88,13 @@ public class HostInterface extends Model {
         rxRate = value;
     }
 
-    private Double rxDrop;
+    private BigInteger rxDrop;
 
-    public Double getRxDrop() {
+    public BigInteger getRxDrop() {
         return rxDrop;
     }
 
-    public void setRxDrop(Double value) {
+    public void setRxDrop(BigInteger value) {
         rxDrop = value;
     }
 
@@ -128,13 +128,13 @@ public class HostInterface extends Model {
         txTotal = value;
     }
 
-    private Double txDrop;
+    private BigInteger txDrop;
 
-    public Double getTxDrop() {
+    public BigInteger getTxDrop() {
         return txDrop;
     }
 
-    public void setTxDrop(Double value) {
+    public void setTxDrop(BigInteger value) {
         txDrop = value;
     }
